tests/test_story4_pytorch.py
#!/usr/bin/env python3
"""
Test suite for Story 4: PyTorch and Core Dependencies Installation
Verifies all acceptance criteria for PyTorch setup.
"""

import os
import subprocess
import unittest
from pathlib import Path


class TestPyTorchInstallation(unittest.TestCase):
    """Test PyTorch installation and CUDA support for MVDream."""
    
    def test_pytorch_installed(self):
        """Test that PyTorch is installed via Poetry."""
        try:
            import torch
            self.assertTrue(True, "PyTorch imported successfully")
        except ImportError:
            self.fail("PyTorch not installed")
    
    def test_torchvision_installed(self):
        """Test that torchvision is installed with compatible version."""
        try:
            import torchvision
            import torch
            
            # Check version compatibility
            torch_version = torch.__version__.split('+')[0]
            torchvision_version = torchvision.__version__.split('+')[0]
            
            # Just verify it's installed - exact version matching is complex
            self.assertTrue(True, f"torchvision {torchvision_version} installed")
            
        except ImportError:
            self.fail("torchvision not installed")
    
    def test_cuda_available(self):
        """Test that torch.cuda.is_available() returns True."""
        import torch
        
        self.assertTrue(torch.cuda.is_available(), 
                       "CUDA is not available in PyTorch")
    
    def test_rtx_3090_device_name(self):
        """Test that torch.cuda.get_device_name(0) returns RTX 3090."""
        import torch
        
        if not torch.cuda.is_available():
            self.skipTest("CUDA not available")
        
        device_name = torch.cuda.get_device_name(0)
        self.assertIn("3090", device_name, 
                     f"RTX 3090 not detected, found: {device_name}")
    
    def test_memory_allocation_20gb(self):
        """Test ability to allocate 20+ GB of GPU memory."""
        import torch
        
        if not torch.cuda.is_available():
            self.skipTest("CUDA not available")
        
        # Clear cache
        torch.cuda.empty_cache()
        
        try:
            # Allocate tensors totaling ~20GB
            allocated = []
            target_gb = 20
            gb_per_tensor = 5
            
            for i in range(target_gb // gb_per_tensor):
                elements = int(gb_per_tensor * (1024**3) / 4)  # float32 = 4 bytes
                height = 50000
                width = elements // height
                tensor = torch.zeros(height, width, device='cuda', dtype=torch.float32)
                allocated.append(tensor)
            
            # Verify total allocation
            total_bytes = sum(t.element_size() * t.nelement() for t in allocated)
            total_gb = total_bytes / (1024**3)
            
            self.assertGreaterEqual(total_gb, 19, 
                                   f"Only allocated {total_gb:.1f}GB, expected 20+GB")
            
            # Cleanup
            del allocated
            torch.cuda.empty_cache()
            
        except RuntimeError as e:
            if "out of memory" in str(e).lower():
                self.fail(f"Could not allocate 20GB of GPU memory: {e}")
            raise
    
    def test_ninja_installed(self):
        """Test that ninja build system is installed."""
        try:
            result = subprocess.run(
                ['ninja', '--version'],
                capture_output=True,
                text=True,
                check=True
            )
            self.assertTrue(result.stdout.strip(), 
                           "Ninja version not found")
        except (subprocess.CalledProcessError, FileNotFoundError):
            self.fail("Ninja build system not installed")
    
    def test_gpu_operations_script(self):
        """Test that test_cuda.py successfully runs GPU operations."""
        test_script = Path.home() / 'mvdream' / 'tests' / 'test_cuda.py'
        
        # Check if script exists
        self.assertTrue(test_script.exists(), 
                       f"GPU test script not found at {test_script}")
        
        # Run the script
        try:
            env = os.environ.copy()
            env['CUDA_HOME'] = '/opt/cuda'
            env['LD_LIBRARY_PATH'] = f"/opt/cuda/lib64:{env.get('LD_LIBRARY_PATH', '')}"
            
            result = subprocess.run(
                ['python', str(test_script)],
                capture_output=True,
                text=True,
                env=env,
                timeout=60
            )
            
            # Check if tests ran (may have some failures due to numerical precision)
            output = result.stderr or result.stdout
            self.assertIn("test_cuda_available", output, 
                         "GPU validation tests did not run properly")
            
        except subprocess.TimeoutExpired:
            self.fail("GPU test script timed out")
        except subprocess.CalledProcessError as e:
            self.fail(f"GPU test script failed: {e.stderr}")
    
    def test_dependencies_locked(self):
        """Test that all dependencies are locked in poetry.lock."""
        lock_file = Path.home() / 'mvdream' / 'poetry.lock'
        
        self.assertTrue(lock_file.exists(), 
                       f"poetry.lock not found at {lock_file}")
        
        # Check that key packages are in lock file
        content = lock_file.read_text()
        
        # These packages should be mentioned in the lock file
        packages = ['torch', 'torchvision', 'numpy', 'pillow']
        
        for package in packages:
            self.assertIn(package, content.lower(), 
                         f"{package} not found in poetry.lock")
    
    def test_readme_updated(self):
        """Test that README.md has PyTorch installation instructions."""
        readme = Path.home() / 'mvdream' / 'README.md'
        
        self.assertTrue(readme.exists(), 
                       f"README.md not found at {readme}")
        
        content = readme.read_text()
        
        # Check for PyTorch installation instructions
        required_sections = [
            'PyTorch',
            'CUDA',
            'pip install torch',
            'test-gpu'
        ]
        
        for section in required_sections:
            self.assertIn(section, content, 
                         f"README missing section about: {section}")
    
    def test_pytorch_cuda_version(self):
        """Test that PyTorch is compiled with CUDA 11.8."""
        import torch
        
        version = torch.__version__
        
        # Check if CUDA version is in the version string
        if '+cu' in version:
            cuda_version = version.split('+cu')[1][:3]
            # Accept 11.8 or 12.x (backward compatible)
            self.assertIn(cuda_version[0:2], ['11', '12'], 
                         f"Unexpected CUDA version: cu{cuda_version}")
    
    def test_numpy_compatibility(self):
        """Test NumPy is installed and compatible with PyTorch."""
        try:
            import numpy as np
            import torch
            
            # Test conversion between NumPy and PyTorch
            np_array = np.array([1.0, 2.0, 3.0])
            torch_tensor = torch.from_numpy(np_array)
            
            self.assertEqual(torch_tensor.shape[0], 3, 
                            "NumPy to PyTorch conversion failed")
            
            # Test reverse conversion
            back_to_numpy = torch_tensor.numpy()
            np.testing.assert_array_equal(np_array, back_to_numpy, 
                                         "PyTorch to NumPy conversion failed")
            
        except ImportError as e:
            self.fail(f"NumPy not installed or incompatible: {e}")
    
    def test_pillow_compatibility(self):
        """Test Pillow is installed for image processing."""
        try:
            from PIL import Image
            import torch
            import torchvision.transforms as transforms
            
            # Create a dummy image
            img = Image.new('RGB', (100, 100), color='red')
            
            # Test torchvision transforms
            transform = transforms.ToTensor()
            tensor = transform(img)
            
            self.assertEqual(tensor.shape, (3, 100, 100), 
                            "Pillow to tensor conversion failed")
            
        except ImportError as e:
            self.fail(f"Pillow not installed or incompatible: {e}")
    
    def test_storage_configuration(self):
        """Test that storage is configured on /mnt/datadrive_m2."""
        # Check if virtual environment is on data drive
        venv_path = Path('/mnt/datadrive_m2/mvdream/.venv')
        self.assertTrue(venv_path.exists(), 
                       f"Virtual environment not on data drive at {venv_path}")
        
        # Check environment variables (if set)
        expected_vars = {
            'PIP_CACHE_DIR': '/mnt/datadrive_m2/.pip-cache',
            'TORCH_HOME': '/mnt/datadrive_m2/.torch',
            'HF_HOME': '/mnt/datadrive_m2/.huggingface'
        }
        
        for var, expected_path in expected_vars.items():
            actual = os.environ.get(var)
            if actual:
                self.assertEqual(actual, expected_path, 
                               f"{var} not set to {expected_path}, found: {actual}")


def run_tests():
    """Run all PyTorch installation tests."""
    loader = unittest.TestLoader()
    suite = loader.loadTestsFromTestCase(TestPyTorchInstallation)
    runner = unittest.TextTestRunner(verbosity=2)
    result = runner.run(suite)
    return result.wasSuccessful()


if __name__ == '__main__':
    success = run_tests()
    exit(0 if success else 1)