tests/test_story5_mvdream.py
#!/usr/bin/env python3
"""
Test suite for Story 5: MVDream Repository Setup
Verifies all acceptance criteria for MVDream integration.
"""

import os
import sys
import unittest
from pathlib import Path


class TestMVDreamRepositorySetup(unittest.TestCase):
    """Test MVDream repository setup and integration."""
    
    def test_mvdream_repository_cloned(self):
        """Test that MVDream repository is cloned to extern/MVDream."""
        mvdream_path = Path('/mnt/datadrive_m2/dream-cad') / 'extern' / 'MVDream'
        self.assertTrue(mvdream_path.exists(), 
                       f"MVDream repository not found at {mvdream_path}")
        
        # Check for key files
        self.assertTrue((mvdream_path / 'setup.py').exists(), 
                       "MVDream setup.py not found")
        self.assertTrue((mvdream_path / 'mvdream').is_dir(), 
                       "MVDream package directory not found")
        self.assertTrue((mvdream_path / 'requirements.txt').exists(), 
                       "MVDream requirements.txt not found")
    
    def test_mvdream_threestudio_cloned(self):
        """Test that MVDream-threestudio is cloned."""
        threestudio_path = Path('/mnt/datadrive_m2/dream-cad') / 'extern' / 'MVDream-threestudio'
        self.assertTrue(threestudio_path.exists(), 
                       f"MVDream-threestudio not found at {threestudio_path}")
        
        # Check for requirements
        self.assertTrue((threestudio_path / 'requirements.txt').exists(), 
                       "MVDream-threestudio requirements.txt not found")
    
    def test_mvdream_installed_editable(self):
        """Test that MVDream is installed as editable package."""
        try:
            import mvdream
            
            # Check if it's installed from the right location
            mvdream_file = Path(mvdream.__file__)
            
            # Check if it's from /mnt/datadrive_m2/dream-cad/extern/MVDream
            expected_in_path = '/dream-cad/extern/MVDream/mvdream'
            
            # The module should be from the extern directory
            self.assertTrue(expected_in_path in str(mvdream_file), 
                           f"MVDream not installed from extern directory: {mvdream_file}")
            
        except ImportError:
            self.fail("MVDream package not installed")
    
    def test_pytorch_lightning_version(self):
        """Test that pytorch-lightning 2.0.9 is installed."""
        try:
            import pytorch_lightning as pl
            
            # Check version
            version = pl.__version__
            # Accept 2.0.9 or 2.0.9.post0 etc
            self.assertTrue(version.startswith('2.0.9'), 
                           f"pytorch-lightning version {version} != 2.0.9")
            
        except ImportError:
            self.fail("pytorch-lightning not installed")
    
    def test_diffusers_installed(self):
        """Test that diffusers library is installed."""
        try:
            import diffusers
            
            # Check that it imports successfully
            self.assertTrue(hasattr(diffusers, '__version__'), 
                           "diffusers missing version attribute")
            
        except ImportError:
            self.fail("diffusers not installed")
    
    def test_transformers_installed(self):
        """Test that transformers library is installed."""
        try:
            import transformers
            
            # Check version
            version = transformers.__version__
            self.assertTrue(version, f"transformers version: {version}")
            
        except ImportError:
            self.fail("transformers not installed")
    
    def test_mvdream_dependencies(self):
        """Test that all MVDream Python dependencies are available."""
        required_packages = [
            'omegaconf',
            'einops',
            'open_clip',
            'opencv-python',
            'imageio',
            'gradio',
            'accelerate',
            'tensorboard',
            'matplotlib',
            'wandb',
            'jaxtyping',
            'typeguard',
            'kornia',
            'sentencepiece',
            'safetensors',
            'huggingface_hub',
            'trimesh',
            'networkx',
            'pysdf',
            'PyMCubes'
        ]
        
        for package in required_packages:
            # Convert package name to import name
            import_name = package.replace('-', '_').lower()
            if import_name == 'opencv_python':
                import_name = 'cv2'
            elif import_name == 'pytorch_lightning':
                import_name = 'pytorch_lightning'
            elif import_name == 'open_clip':
                import_name = 'open_clip'
            elif import_name == 'huggingface_hub':
                import_name = 'huggingface_hub'
            elif import_name == 'pymcubes':
                import_name = 'mcubes'
            
            try:
                __import__(import_name)
            except ImportError:
                self.fail(f"Required package {package} (import as {import_name}) not installed")
    
    def test_project_structure_documented(self):
        """Test that project structure is documented."""
        doc_path = Path('/mnt/datadrive_m2/dream-cad') / 'docs' / 'project-structure.md'
        self.assertTrue(doc_path.exists(), 
                       f"Project structure documentation not found at {doc_path}")
        
        # Check content
        content = doc_path.read_text()
        required_sections = [
            'Directory Tree',
            'extern/MVDream',
            'MVDream-threestudio',
            'Installed Packages',
            'Environment Variables'
        ]
        
        for section in required_sections:
            self.assertIn(section, content, 
                         f"Documentation missing section: {section}")
    
    def test_mvdream_imports(self):
        """Test that MVDream modules can be imported."""
        try:
            # These imports should work if MVDream is properly installed
            import mvdream
            from mvdream import camera_utils
            from mvdream import model_zoo
            from mvdream.ldm import models
            
            # Check that the package has expected attributes
            self.assertTrue(hasattr(mvdream, '__file__'), 
                           "MVDream package missing expected attributes")
            
        except ImportError as e:
            self.fail(f"Failed to import MVDream modules: {e}")
    
    def test_open_clip_version(self):
        """Test that open-clip-torch 2.7.0 is installed."""
        try:
            import open_clip
            
            # The version check might not work directly, but import should succeed
            self.assertTrue(True, "open-clip-torch imported successfully")
            
        except ImportError:
            self.fail("open-clip-torch not installed")
    
    def test_omegaconf_version(self):
        """Test that omegaconf 2.3.0 is installed."""
        try:
            import omegaconf
            
            version = omegaconf.__version__
            self.assertEqual(version, '2.3.0', 
                           f"omegaconf version {version} != 2.3.0")
            
        except ImportError:
            self.fail("omegaconf not installed")
    
    def test_git_repositories_valid(self):
        """Test that cloned repositories are valid git repos."""
        repos = [
            Path('/mnt/datadrive_m2/dream-cad') / 'extern' / 'MVDream',
            Path('/mnt/datadrive_m2/dream-cad') / 'extern' / 'MVDream-threestudio'
        ]
        
        for repo in repos:
            git_dir = repo / '.git'
            self.assertTrue(git_dir.exists(), 
                           f"Git directory not found in {repo}")
    
    def test_pyproject_toml_updated(self):
        """Test that pyproject.toml contains MVDream dependencies."""
        pyproject_path = Path('/mnt/datadrive_m2/dream-cad') / 'pyproject.toml'
        content = pyproject_path.read_text()
        
        # Check for key dependencies
        dependencies = [
            'pytorch-lightning',
            'diffusers',
            'transformers',
            'omegaconf',
            'einops',
            'opencv-python',
            'imageio',
            'gradio'
        ]
        
        for dep in dependencies:
            self.assertIn(dep, content, 
                         f"Dependency {dep} not found in pyproject.toml")
    
    def test_submodules_configured(self):
        """Test that git submodules are properly configured."""
        gitmodules_path = Path('/mnt/datadrive_m2/dream-cad') / '.gitmodules'
        self.assertTrue(gitmodules_path.exists(), 
                       ".gitmodules file not found")
        
        content = gitmodules_path.read_text()
        
        # Check for both submodules
        self.assertIn('extern/MVDream', content, 
                     "MVDream submodule not configured")
        self.assertIn('extern/MVDream-threestudio', content, 
                     "MVDream-threestudio submodule not configured")
        self.assertIn('https://github.com/bytedance/MVDream.git', content, 
                     "MVDream URL not found in .gitmodules")
        self.assertIn('https://github.com/bytedance/MVDream-threestudio.git', content, 
                     "MVDream-threestudio URL not found in .gitmodules")


def run_tests():
    """Run all MVDream repository setup tests."""
    loader = unittest.TestLoader()
    suite = loader.loadTestsFromTestCase(TestMVDreamRepositorySetup)
    runner = unittest.TextTestRunner(verbosity=2)
    result = runner.run(suite)
    return result.wasSuccessful()


if __name__ == '__main__':
    success = run_tests()
    sys.exit(0 if success else 1)